+++ src/mappingForSingleRandomWinnerProxyFactory.ts
@@ -1,6 +1,6 @@
 import {
   ProxyCreated, // note that the version doesn't matter here, it's just the way the generated code works
-} from '../generated/SingleRandomWinnerProxyFactory_v3_1_0/SingleRandomWinnerProxyFactory'
+} from '../generated/SingleRandomWinnerProxyFactory/SingleRandomWinnerProxyFactory'
 
 import {
   SingleRandomWinner as SingleRandomWinnerTemplate,
