+++ src/mappingForStakePrizePoolProxyFactory.ts
@@ -1,16 +1,14 @@
 import {
     ProxyCreated,
-  } from '../generated/StakePrizePoolProxyFactory_v3_1_0/StakePrizePoolProxyFactory'
+  } from '../generated/StakePrizePoolProxyFactory/StakePrizePoolProxyFactory'
   
   import {
-    PrizePool_v3 as PrizePoolV3Template,
-    PrizePool_v3_1_0 as PrizePoolTemplate,
+    PrizePool as PrizePoolTemplate,
     StakePrizePool as StakePrizeTemplate,
   } from '../generated/templates'
   
   export function handleStakePrizePoolCreated(event: ProxyCreated): void {
     // Start listening for events from the dynamically generated contract
-    PrizePoolV3Template.create(event.params.proxy)
     PrizePoolTemplate.create(event.params.proxy)
     StakePrizeTemplate.create(event.params.proxy)
   }
