+++ src/mappingForYieldSourcePrizePoolProxyFactory.ts
@@ -1,15 +1,15 @@
 import {
   ProxyCreated,
-} from '../generated/CompoundPrizePoolProxyFactory_v3_0_1/CompoundPrizePoolProxyFactory_v3_0_1'
+} from "../generated/YieldSourcePrizePoolProxyFactory/YieldSourcePrizePoolProxyFactory"
 
 import {
-  PrizePool_v3_1_0 as PrizePoolTemplate,
-  CompoundPrizePool
+
+  PrizePool as PrizePoolTemplate,
+  YieldSourcePrizePool as YieldSourcePrizePoolTemplate,
 } from '../generated/templates'
 
 export function handleProxyCreated(event: ProxyCreated): void {
   // Start listening for events from the dynamically generated contract
   PrizePoolTemplate.create(event.params.proxy)
-  CompoundPrizePool.create(event.params.proxy)
-
+  YieldSourcePrizePoolTemplate.create(event.params.proxy)
 }
